Changed the UI flow around showing and creating accounts. Added an 'add account' screen.
Organized the ui so that when there is only one account, the software continues to operate similar to how it works now. The wallet list is only shown when there are multiple accounts.

Fix back button navigation.
Use different template for add -vs- update for accountConfig.
Move the device config button to the left.

Bump to 0.8.10